+++ userprofile/tests/test_views.py
@@ -1,7 +1,7 @@
 """Tests for views in the userprofile Django app"""
 
-from django.core.urlresolvers import reverse
 from django.test import RequestFactory
+from django.urls import reverse
 
 from test_plus.test import TestCase